+++ src/al/alWriter.ts
@@ -1,15 +1,9 @@
-import * as vscode from 'vscode';
-import * as fs from 'fs';
 import { Position, TextEditor } from 'vscode';
 
 export class ALWriter {
     private indentText : string;
     private indentPart : string;
     public content : string;
-    public targetALFileName: string = "";
-    public targetALObjectName: string = "";
-    public targetALProcedureName: string = "";
-    public targetALFilePath: string = "";
 
     constructor() {
         this.content = "";
@@ -26,26 +20,6 @@ export class ALWriter {
         await editor.edit(editBuilder => {
             editBuilder.insert(new Position(lineNo, 0), content);
         });
-    }
-
-    writeStubInFile(stub: string): void {
-        let fileText : string = fs.readFileSync(this.targetALFilePath, 'utf-8');
-        let lastIndexClosingBracket = fileText.lastIndexOf("}");
-        if (lastIndexClosingBracket === -1) {
-            console.error("The file seems to be not properly built. Could not find closing bracket.");
-        }
-        else {
-            let textBeforeLastBracket : string = fileText.substring(0, lastIndexClosingBracket);
-            let textAfterLastBracket : string = fileText.substring(lastIndexClosingBracket + 1);
-            let fileTextWithStub : string = textBeforeLastBracket + "\n" + "   " + stub + "\n" + "}";
-            if (textAfterLastBracket) {
-                fileTextWithStub += textAfterLastBracket;
-            }
-            fs.writeFileSync(this.targetALFilePath, fileTextWithStub, 'utf-8');
-            //this._alFiles.repopulateALFIlesArray();
-            console.log('Stub added.');
-        }
-        
     }
 
         public toString() : string {
@@ -67,10 +41,6 @@ export class ALWriter {
 
         public writeLine(line : string) {
             this.content += (this.indentText + line + "\n");
-        }
-
-        public writeProcedureStub(procedureStub: string) {
-            this.writeLine(procedureStub);
         }
 
         public clearContent() {